requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cython==3.0.12
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
ftfy==6.3.1
grpcio==1.70.0
huggingface-hub==0.30.2
idna==3.4
imageio==2.37.0
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.8
lazy-loader==0.4
markdown==3.7
markupsafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.2.0
mpmath==1.3.0
muon-optimizer==0.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.23.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.0.1
pipreqs==0.5.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycocotools==2.0.8
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.37.0
regex==2024.11.6
requests==2.28.1
rpds-py==0.30.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==2.1.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==1.26.13
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.22.post4
yacs==0.1.8
yapf==0.43.0
yarg==0.1.9
zipp==3.21.0
